# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CorpoPay API",
    "version": "0.1.0",
    "description": "Multi-tenant payment orchestration API."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "PaymentLinkCreateResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ]
          },
          "currency": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "isRecurring": {
            "type": "boolean"
          },
          "billingInterval": {
            "type": "string",
            "nullable": true
          },
          "intervalValue": {
            "type": "number",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "url",
          "amount",
          "currency",
          "description",
          "reference",
          "status",
          "isRecurring",
          "billingInterval",
          "intervalValue",
          "createdAt"
        ]
      },
      "PaymentLinkListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ]
          },
          "currency": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "attemptCount": {
            "type": "number"
          },
          "maxAttempts": {
            "type": "number"
          },
          "expiresAt": {
            "type": "string",
            "nullable": true
          },
          "isRecurring": {
            "type": "boolean"
          },
          "billingInterval": {
            "type": "string",
            "nullable": true
          },
          "intervalValue": {
            "type": "number",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "url",
          "amount",
          "currency",
          "description",
          "reference",
          "provider",
          "status",
          "attemptCount",
          "maxAttempts",
          "expiresAt",
          "isRecurring",
          "billingInterval",
          "intervalValue",
          "createdAt"
        ]
      },
      "PaymentLinkDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ]
          },
          "currency": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "customerEmail": {
            "type": "string",
            "nullable": true
          },
          "customerPhone": {
            "type": "string",
            "nullable": true
          },
          "provider": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "maxAttempts": {
            "type": "number"
          },
          "attemptCount": {
            "type": "number"
          },
          "expiresAt": {
            "type": "string",
            "nullable": true
          },
          "isRecurring": {
            "type": "boolean"
          },
          "billingInterval": {
            "type": "string",
            "nullable": true
          },
          "intervalValue": {
            "type": "number",
            "nullable": true
          },
          "maxRetries": {
            "type": "number"
          },
          "isInstallment": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "_count": {
            "type": "object",
            "properties": {
              "paymentIntents": {
                "type": "number"
              }
            },
            "required": [
              "paymentIntents"
            ]
          }
        },
        "required": [
          "id",
          "tenantId",
          "slug",
          "amount",
          "currency",
          "description",
          "reference",
          "customerName",
          "customerEmail",
          "customerPhone",
          "provider",
          "status",
          "maxAttempts",
          "attemptCount",
          "expiresAt",
          "isRecurring",
          "billingInterval",
          "intervalValue",
          "maxRetries",
          "isInstallment",
          "createdAt",
          "updatedAt",
          "url",
          "_count"
        ]
      },
...
      "PayRequest": {
        "type": "object",
        "properties": {
          "customerIp": {
            "type": "string"
          },
          "customerEmail": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "string",
                "enum": [
                  ""
                ]
              }
            ]
          },
          "customerName": {
            "type": "string"
          },
          "installmentPlanId": {
            "type": "string"
          },
          "downPaymentAmount": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": true
          }
        }
      },
      "PayResponse": {
        "type": "object",
        "properties": {
          "intentId": {
            "type": "string"
          },
          "agreementId": {
            "type": "string"
          },
          "redirectUrl": {
            "type": "string"
          },
          "providerData": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {
              "nullable": true
            }
          }
        },
        "required": [
          "intentId",
          "redirectUrl",
          "providerData"
        ]
      },
      "PublicRelayResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "providerData": {
            "type": "object",
            "nullable": true,
            "properties": {
              "paywallUrl": {
                "type": "string"
              },
              "payload": {
                "type": "string"
              },
              "signature": {
                "type": "string"
              },
              "mode": {
                "type": "string"
              },
              "chargeId": {
                "type": "string"
              },
              "redirectUrl": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "status",
          "providerData"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/public/checkout/{slug}/pay": {
      "post": {
        "operationId": "publicPay",
        "summary": "Create a payment intent from a payment link",
        "tags": [
          "Public"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "slug",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayResponse"
                }
              }
            }
          },
          "400": {
            "description": "Plan required / amount constraints"
          },
          "404": {
            "description": "Payment link not found"
          },
          "410": {
            "description": "Link inactive / expired"
          },
          "429": {
            "description": "Maximum payment attempts reached"
          }
        }
      }
    },
...
    "/public/pay/{correlationId}": {
      "get": {
        "operationId": "publicRelay",
        "summary": "Fetch persisted provider data for the hosted relay page",
        "tags": [
          "Public"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "correlationId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK (JSON for API clients; HTML for browsers)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicRelayResponse"
                }
              }
            }
          },
          "404": {
            "description": "Payment session not found"
          }
        }
      }
    },
...
    "/payment-links": {
      "post": {
        "operationId": "createPaymentLink",
        "summary": "Create a payment link",
        "tags": [
          "Payment Links"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number",
                    "minimum": 0,
                    "exclusiveMinimum": true,
                    "maximum": 100000000
                  },
                  "currency": {
                    "type": "string",
                    "default": "MAD"
                  },
                  "description": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "reference": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "provider": {
                    "type": "string",
                    "enum": [
                      "NAPS",
                      "VPS",
                      "STRIPE"
                    ]
                  },
                  "customerName": {
                    "type": "string",
                    "maxLength": 100
                  },
                  "customerEmail": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "email"
                      },
                      {
                        "type": "string",
                        "enum": [
                          ""
                        ]
                      }
                    ]
                  },
                  "customerPhone": {
                    "type": "string",
                    "maxLength": 20
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "maxAttempts": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10,
                    "default": 1
                  },
                  "isRecurring": {
                    "type": "boolean",
                    "default": false
                  },
                  "billingInterval": {
                    "type": "string",
                    "enum": [
                      "DAILY",
                      "WEEKLY",
                      "MONTHLY",
                      "QUARTERLY",
                      "ANNUAL",
                      "CUSTOM"
                    ]
                  },
                  "intervalValue": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 365,
                    "default": 1
                  },
                  "maxRetries": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10,
                    "default": 3
                  },
                  "isInstallment": {
                    "type": "boolean",
                    "default": false
                  }
                },
                "required": [
                  "amount",
                  "description",
                  "reference",
                  "provider"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentLinkCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Provider not configured"
          }
        }
      },
      "get": {
        "operationId": "listPaymentLinks",
        "summary": "List payment links",
        "tags": [
          "Payment Links"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "offset",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "status",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PaymentLinkListItem"
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "page": {
                      "type": "number"
                    },
                    "limit": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "data",
                    "total",
                    "page",
                    "limit"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/payment-links/{id}": {
      "get": {
        "operationId": "getPaymentLink",
        "summary": "Get a payment link",
        "tags": [
          "Payment Links"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentLinkDetail"
                }
              }
            }
          },
          "404": {
            "description": "Payment link not found"
          }
        }
      }
    },
    "/payment-links/{id}/cancel": {
      "patch": {
        "operationId": "cancelPaymentLink",
        "summary": "Cancel a payment link",
        "tags": [
          "Payment Links"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "status"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Link not active"
          },
          "404": {
            "description": "Payment link not found"
          }
        }
      }
    },
...
  }
}